Fixed up how the form works
