+++ src/types/forms/donation.ts
@@ -2,12 +2,6 @@ import { z } from 'zod';
 
 export const zDonationFormData = z.object({
   donationDate: z.date(),
-  category: z.string().min(1, { message: 'Required' }),
-  donatedItemName: z.string().min(1, { message: 'Required' }),
-  quantity: z.number().positive().min(1, { message: 'Required' }),
-  newOrUsed: z.string().min(1, { message: 'Required' }),
-  highOrLow: z.string().optional(),
-  price: z.number().positive().min(1, { message: 'Required' }),
   user: z.string().min(1, { message: 'Required' }),
 });
